src/main/java/racingcar/model/Cars.java
package racingcar.model;

import java.util.List;

public class Cars {

    private final List<Car> cars;

    public Cars(List<Car> cars) {
        this.cars = cars;
    }

    public void moveEligibleCars(MoveStrategy moveStrategy) {
        cars.stream()
                .filter(car -> moveStrategy.canMove())
                .forEach(Car::moveForward);
    }

    public List<Integer> getPositions() {
        return cars.stream()
                .map(Car::getPosition)
                .toList();
    }

    public List<String> getNames() {
        return cars.stream()
                .map(Car::getName)
                .toList();
    }

}
